feat:add countries to store and disable submit button
