--- src/constants/constants.ts
@@ -1,290 +1,0 @@
-import * as yup from 'yup';
-
-export const schema = yup.object().shape({
-  name: yup
-    .string()
-    .required('Name is a required field')
-    .matches(/^[A-Z]/, 'Name must start with an uppercase letter'),
-  age: yup
-    .number()
-    .required('Age is a required field')
-    .positive('Age must be a positive number')
-    .integer('Age must be an integer'),
-  gender: yup
-    .string()
-    .oneOf(['male', 'female'])
-    .required('Choose the gender, this is required'),
-  country: yup.string().required('Country is a required field'),
-  picture: yup
-    .mixed()
-    .required('Image is a required field')
-    .test(
-      'fileType',
-      'Invalid file type, only JPEG and PNG are allowed',
-      (value) => {
-        const file = value as File;
-        return !file || ['image/jpeg', 'image/png'].includes(file.type);
-      }
-    )
-    .test('fileSize', 'File size is too large', (value) => {
-      const file = value as File;
-      return !file || file.size <= 1024 * 1024 * 2;
-    }),
-  email: yup.string().email().required(),
-  password: yup
-    .string()
-    .required('Password is required')
-    .matches(
-      /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[\W_])/,
-      'Password must contain at least one number, one uppercase letter, one lowercase letter, and one special character'
-    ),
-  confirmPassword: yup
-    .string()
-    .oneOf([yup.ref('password')], 'Passwords must match')
-    .required('Please confirm your password'),
-  accept: yup.boolean().test({
-    name: 'accepted',
-    message: 'You must accept the terms and conditions',
-    test: (value) => value === true,
-  }),
-});
-
-export const countries = [
-  'Afghanistan',
-  'Albania',
-  'Algeria',
-  'American Samoa',
-  'Andorra',
-  'Angola',
-  'Anguilla',
-  'Antigua and Barbuda',
-  'Argentina',
-  'Armenia',
-  'Aruba',
-  'Australia',
-  'Austria',
-  'Azerbaijan',
-  'Bangladesh',
-  'Barbados',
-  'Bahamas',
-  'Bahrain',
-  'Belarus',
-  'Belgium',
-  'Belize',
-  'Benin',
-  'Bermuda',
-  'Bhutan',
-  'Bolivia',
-  'Bosnia and Herzegovina',
-  'Botswana',
-  'Brazil',
-  'British Indian Ocean Territory',
-  'British Virgin Islands',
-  'Brunei Darussalam',
-  'Bulgaria',
-  'Burkina Faso',
-  'Burma',
-  'Burundi',
-  'Cambodia',
-  'Cameroon',
-  'Canada',
-  'Cape Verde',
-  'Cayman Islands',
-  'Central African Republic',
-  'Chad',
-  'Chile',
-  'China',
-  'Christmas Island',
-  'Cocos (Keeling) Islands',
-  'Colombia',
-  'Comoros',
-  'Congo-Brazzaville',
-  'Congo-Kinshasa',
-  'Cook Islands',
-  'Costa Rica',
-  'Croatia',
-  'Cura?ao',
-  'Cyprus',
-  'Czech Republic',
-  'Denmark',
-  'Djibouti',
-  'Dominica',
-  'Dominican Republic',
-  'East Timor',
-  'Ecuador',
-  'El Salvador',
-  'Egypt',
-  'Equatorial Guinea',
-  'Eritrea',
-  'Estonia',
-  'Ethiopia',
-  'Falkland Islands',
-  'Faroe Islands',
-  'Federated States of Micronesia',
-  'Fiji',
-  'Finland',
-  'France',
-  'French Guiana',
-  'French Polynesia',
-  'French Southern Lands',
-  'Gabon',
-  'Gambia',
-  'Georgia',
-  'Germany',
-  'Ghana',
-  'Gibraltar',
-  'Greece',
-  'Greenland',
-  'Grenada',
-  'Guadeloupe',
-  'Guam',
-  'Guatemala',
-  'Guernsey',
-  'Guinea',
-  'Guinea-Bissau',
-  'Guyana',
-  'Haiti',
-  'Heard and McDonald Islands',
-  'Honduras',
-  'Hong Kong',
-  'Hungary',
-  'Iceland',
-  'India',
-  'Indonesia',
-  'Iraq',
-  'Ireland',
-  'Isle of Man',
-  'Israel',
-  'Italy',
-  'Jamaica',
-  'Japan',
-  'Jersey',
-  'Jordan',
-  'Kazakhstan',
-  'Kenya',
-  'Kiribati',
-  'Kuwait',
-  'Kyrgyzstan',
-  'Laos',
-  'Latvia',
-  'Lebanon',
-  'Lesotho',
-  'Liberia',
-  'Libya',
-  'Liechtenstein',
-  'Lithuania',
-  'Luxembourg',
-  'Macau',
-  'Macedonia',
-  'Madagascar',
-  'Malawi',
-  'Malaysia',
-  'Maldives',
-  'Mali',
-  'Malta',
-  'Marshall Islands',
-  'Martinique',
-  'Mauritania',
-  'Mauritius',
-  'Mayotte',
-  'Mexico',
-  'Moldova',
-  'Monaco',
-  'Mongolia',
-  'Montenegro',
-  'Montserrat',
-  'Morocco',
-  'Mozambique',
-  'Namibia',
-  'Nauru',
-  'Nepal',
-  'Netherlands',
-  'New Caledonia',
-  'New Zealand',
-  'Nicaragua',
-  'Niger',
-  'Nigeria',
-  'Niue',
-  'Norfolk Island',
-  'Northern Mariana Islands',
-  'Norway',
-  'Oman',
-  'Pakistan',
-  'Palau',
-  'Panama',
-  'Papua New Guinea',
-  'Paraguay',
-  'Peru',
-  'Philippines',
-  'Pitcairn Islands',
-  'Poland',
-  'Portugal',
-  'Puerto Rico',
-  'Qatar',
-  'R?union',
-  'Romania',
-  'Russia',
-  'Rwanda',
-  'Saint Barth?lemy',
-  'Saint Helena',
-  'Saint Kitts and Nevis',
-  'Saint Lucia',
-  'Saint Martin',
-  'Saint Pierre and Miquelon',
-  'Saint Vincent',
-  'Samoa',
-  'San Marino',
-  'S?o Tom? and Pr?ncipe',
-  'Saudi Arabia',
-  'Senegal',
-  'Serbia',
-  'Seychelles',
-  'Sierra Leone',
-  'Singapore',
-  'Sint Maarten',
-  'Slovakia',
-  'Slovenia',
-  'Solomon Islands',
-  'Somalia',
-  'South Africa',
-  'South Georgia',
-  'South Korea',
-  'Spain',
-  'Sri Lanka',
-  'Sudan',
-  'Suriname',
-  'Svalbard and Jan Mayen',
-  'Sweden',
-  'Swaziland',
-  'Switzerland',
-  'Syria',
-  'Taiwan',
-  'Tajikistan',
-  'Tanzania',
-  'Thailand',
-  'Togo',
-  'Tokelau',
-  'Tonga',
-  'Trinidad and Tobago',
-  'Tunisia',
-  'Turkey',
-  'Turkmenistan',
-  'Turks and Caicos Islands',
-  'Tuvalu',
-  'Uganda',
-  'Ukraine',
-  'United Arab Emirates',
-  'United Kingdom',
-  'United States',
-  'Uruguay',
-  'Uzbekistan',
-  'Vanuatu',
-  'Vatican City',
-  'Vietnam',
-  'Venezuela',
-  'Wallis and Futuna',
-  'Western Sahara',
-  'Yemen',
-  'Zambia',
-  'Zimbabwe',
-];
