+++ src/pages/UncontrolledForm.tsx
@@ -1,7 +1,6 @@
 import { useRef, useState } from 'react';
-import { countries } from '../constants/constants';
 import styles from '../styles/form.module.css';
-import { schema } from '../constants/constants';
+import { schema } from '../utils/yup';
 import { FormField, IUncontrolledForm } from '../types/types';
 import { ValidationError } from 'yup';
 import { convertImage } from '../utils/utils';
@@ -25,6 +24,7 @@ const UncontrolledForm = () => {
   const actualData = useAppSelector((store) => store.data);
 
   const [errors, setErrors] = useState<IUncontrolledForm>({});
+  const countries = useAppSelector((store) => store.countries);
   const [ShowImage, setShowImage] = useState<string>('');
   const [imageObject, setImageObject] = useState<File>();
   const dispatch = useAppDispatch();
